Fixing missing tooltip on status page
