+++ web/locales/en/plugin__netobserv-plugin.json
@@ -211,7 +211,6 @@
   "Cancel": "Cancel",
   "Delete": "Delete",
   "Network Observability FlowCollector status": "Network Observability FlowCollector status",
-  "FlowCollector status": "FlowCollector status",
   "Configuration error": "Configuration error",
   "Configuration warnings": "Configuration warnings",
   "Network Observability is on hold": "Network Observability is on hold",
@@ -528,6 +527,7 @@
   "FlowCollector has errors": "FlowCollector has errors",
   "FlowCollector is on hold": "FlowCollector is on hold",
   "Loading FlowCollector status...": "Loading FlowCollector status...",
+  "FlowCollector status": "FlowCollector status",
   "Show all graphs": "Show all graphs",
   "Focus on this graph": "Focus on this graph",
   "Show total": "Show total",
